change pages return policy and oferta
